+++ api/database/migrations/2019_01_04_123733_create_peoples_table.php
@@ -14,18 +14,18 @@ class CreatePeoplesTable extends Migration
     public function up()
     {
         Schema::create('peoples', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title');
             $table->text('description');
-            $table->integer('type_id')->index()->unsigned();
-            $table->integer('organisation_id')->index()->unsigned();
+            $table->bigInteger('type_id')->index()->unsigned();
+            $table->bigInteger('organisation_id')->index()->unsigned();
             $table->text('icon_path');
             $table->text('icon_preset');
             $table->double('position_x');
             $table->double('position_y');
             $table->tinyInteger('trajectory')->default(1);
-            $table->integer('character_id')->unsigned()->index();
-            $table->integer('parent_id')->unsigned()->index();
+            $table->bigInteger('character_id')->unsigned()->index();
+            $table->bigInteger('parent_id')->unsigned()->index();
             $table->tinyInteger('status');
             $table->timestamps();
         });
